move lpa tracker to page
